chore: move jobs files to one module
