+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/available-jobs-data.ts
@@ -5,9 +5,8 @@ import { useParams } from 'react-router-dom';
 import { apiClient } from '@/api/api-client';
 import { apiPaths } from '@/api/api-paths';
 import { stringifyUrlQueryObject } from '@/shared/helpers/transfomers';
-import type { JobsFilterStoreProps } from '@/modules/worker/hooks/use-jobs-filter-store';
-import { useJobsFilterStore } from '@/modules/worker/hooks/use-jobs-filter-store';
 import { createPaginationSchema } from '@/shared/helpers/pagination';
+import { type JobsFilterStoreProps, useJobsFilterStore } from '../hooks';
 
 const availableJobSchema = z.object({
   escrow_address: z.string(),
