+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/components/desktop/available-jobs-table-desktop.tsx
@@ -4,13 +4,13 @@ import {
 } from 'material-react-table';
 import { t } from 'i18next';
 import { useEffect, useMemo } from 'react';
-import { useJobsFilterStore } from '@/modules/worker/hooks/use-jobs-filter-store';
 import { usePagination } from '@/modules/worker/hooks/use-pagination';
-import { useGetAvailableJobsData } from '@/modules/worker/services/available-jobs-data';
 import { createTableDarkMode } from '@/shared/styles/create-table-dark-mode';
-import { EscrowAddressSearchForm } from '@/modules/worker/jobs/escrow-address-search-form';
 import { useGetAvailableJobsColumns } from '@/modules/worker/jobs/available-jobs/hooks/use-get-available-jobs-columns';
 import { useColorMode } from '@/shared/contexts/color-mode';
+import { useJobsFilterStore } from '../../../hooks';
+import { useGetAvailableJobsData } from '../../available-jobs-data';
+import { EscrowAddressSearchForm } from '../../../components';
 
 interface AvailableJobsTableProps {
   chainIdsEnabled: number[];
