+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/components/mobile/available-jobs-table-mobile.tsx
@@ -1,9 +1,9 @@
 import { useTranslation } from 'react-i18next';
 import { Button } from '@/shared/components/ui/button';
 import { FiltersButtonIcon } from '@/shared/components/ui/icons';
-import { useJobsFilterStore } from '@/modules/worker/hooks/use-jobs-filter-store';
-import { EscrowAddressSearchForm } from '@/modules/worker/jobs/escrow-address-search-form';
 import { AvailableJobsTableJobsListMobile } from '@/modules/worker/jobs/available-jobs/components/mobile/available-jobs-table-jobs-list-mobile';
+import { useJobsFilterStore } from '../../../hooks';
+import { EscrowAddressSearchForm } from '../../../components';
 
 interface AvailableJobsTableMobileProps {
   handleOpenMobileFilterDrawer: () => void;
