+++ packages/apps/human-app/frontend/src/modules/worker/jobs/available-jobs/hooks/use-get-available-jobs-columns.tsx
@@ -3,19 +3,18 @@ import type { MRT_ColumnDef } from 'material-react-table';
 import { t } from 'i18next';
 import { Grid } from '@mui/material';
 import { useMemo } from 'react';
-import { type AvailableJob } from '@/modules/worker/services/available-jobs-data';
-import { useAssignJobMutation } from '@/modules/worker/services/use-assign-job';
-import { EvmAddress } from '@/modules/worker/jobs/evm-address';
-import { RewardAmount } from '@/modules/worker/jobs/reward-amount';
+import { useAssignJobMutation } from '@/modules/worker/jobs/available-jobs/hooks/use-assign-job';
 import { getNetworkName } from '@/modules/smart-contracts/get-network-name';
 import { Chip } from '@/shared/components/ui/chip';
-import { useJobsNotifications } from '@/modules/worker/hooks/use-jobs-notifications';
 import { TableButton } from '@/shared/components/ui/table-button';
 import { TableHeaderCell } from '@/shared/components/ui/table/table-header-cell';
 import { AvailableJobsNetworkFilter } from '@/modules/worker/jobs/available-jobs/components/available-jobs-network-filter';
 import { AvailableJobsRewardAmountSort } from '@/modules/worker/jobs/available-jobs/components/available-jobs-reward-amount-sort';
 import { AvailableJobsJobTypeFilter } from '@/modules/worker/jobs/available-jobs/components/available-jobs-job-type-filter';
 import type { JobType } from '@/modules/smart-contracts/EthKVStore/config';
+import { useJobsNotifications } from '../../hooks';
+import { type AvailableJob } from '../available-jobs-data';
+import { EvmAddress, RewardAmount } from '../../components';
 
 const COL_SIZE = 100;
 const COL_SIZE_LG = 200;
