+++ packages/apps/human-app/frontend/src/modules/worker/jobs/components/reward-amount.tsx
