+++ packages/apps/human-app/frontend/src/modules/worker/jobs/hooks/use-get-all-networks.ts
