+++ packages/apps/human-app/frontend/src/modules/worker/jobs/hooks/use-my-jobs-filter-store.tsx
@@ -1,51 +1,48 @@
 /* eslint-disable camelcase -- api params*/
 import { create } from 'zustand';
 import type { PageSize } from '@/shared/types/entity.type';
-import { MyJobStatus } from '@/modules/worker/services/my-jobs-data';
+import { type MyJobStatus } from '../types';
 
-export interface JobsFilterStoreProps {
+export interface MyJobsFilterStoreProps {
   filterParams: {
     sort?: 'asc' | 'desc';
     sort_field?:
       | 'chain_id'
       | 'job_type'
       | 'reward_amount'
-      | 'created_at'
-      | 'escrow_address';
-    network?: 'MATIC' | 'POLYGON';
-    // TODO add allowed job types
+      | 'expires_at'
+      | 'created_at';
     job_type?: string;
     status?: MyJobStatus;
     escrow_address?: string;
     page: number;
     page_size: PageSize;
-    fields: string[];
     chain_id?: number;
   };
+  availableJobTypes: string[];
   setFilterParams: (
-    partialParams: Partial<JobsFilterStoreProps['filterParams']>
+    partialParams: Partial<MyJobsFilterStoreProps['filterParams']>
   ) => void;
   resetFilterParams: () => void;
   setSearchEscrowAddress: (escrow_address: string) => void;
   setOracleAddress: (oracleAddress: string) => void;
+  setAvailableJobTypes: (jobTypes: string[]) => void;
   setPageParams: (pageIndex: number, pageSize: PageSize) => void;
 }
 
 const initialFiltersState = {
+  escrow_address: '',
   page: 0,
   page_size: 5,
-  fields: ['reward_amount', 'job_description', 'reward_token'],
-  status: MyJobStatus.ACTIVE,
-  escrow_address: '',
-} satisfies Pick<
-  JobsFilterStoreProps['filterParams'],
-  'page_size' | 'page' | 'fields' | 'status' | 'escrow_address'
->;
+  sort_field: 'created_at',
+  sort: 'desc',
+} as const;
 
-export const useJobsFilterStore = create<JobsFilterStoreProps>((set) => ({
+export const useMyJobsFilterStore = create<MyJobsFilterStoreProps>((set) => ({
   filterParams: initialFiltersState,
+  availableJobTypes: [],
   setFilterParams: (
-    partialParams: Partial<JobsFilterStoreProps['filterParams']>
+    partialParams: Partial<MyJobsFilterStoreProps['filterParams']>
   ) => {
     set((state) => ({
       ...state,
@@ -83,8 +80,14 @@ export const useJobsFilterStore = create<JobsFilterStoreProps>((set) => ({
       ...state,
       filterParams: {
         ...state.filterParams,
-        oracle_address: oracleAddress,
+        address: oracleAddress,
       },
+    }));
+  },
+  setAvailableJobTypes: (jobTypes: string[]) => {
+    set((state) => ({
+      ...state,
+      jobTypes,
     }));
   },
 }));
