+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/desktop/my-jobs-job-type-filter.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable camelcase --- ... */
 import { useTranslation } from 'react-i18next';
-import { useMyJobsFilterStore } from '@/modules/worker/hooks/use-my-jobs-filter-store';
 import { Filtering } from '@/shared/components/ui/table/table-header-menu.tsx/filtering';
 import { JOB_TYPES } from '@/shared/consts';
+import { useMyJobsFilterStore } from '../../hooks';
 
 export function MyJobsJobTypeFilter() {
   const { t } = useTranslation();
