+++ packages/apps/human-app/frontend/src/modules/worker/jobs/my-jobs/status-chip.tsx
@@ -1,11 +1,11 @@
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import { type MyJob } from '@/modules/worker/services/my-jobs-data';
 import { colorPalette as lightModeColorPalette } from '@/shared/styles/color-palette';
-import { getChipStatusColor } from '@/modules/worker/utils/get-chip-status-color';
 import { useColorMode } from '@/shared/contexts/color-mode';
+import { type MyJob } from '../components';
+import { getChipStatusColor } from './utils/get-chip-status-color';
 
-export function StatusChip({ status }: { status: MyJob['status'] }) {
+export function StatusChip({ status }: Readonly<{ status: MyJob['status'] }>) {
   const { colorPalette } = useColorMode();
 
   return (
