+++ packages/apps/human-app/frontend/src/modules/worker/jobs/types.ts
@@ -1,0 +1,10 @@
+export enum MyJobStatus {
+  ACTIVE = 'ACTIVE',
+  CANCELED = 'CANCELED',
+  COMPLETED = 'COMPLETED',
+  VALIDATION = 'VALIDATION',
+  EXPIRED = 'EXPIRED',
+  REJECTED = 'REJECTED',
+}
+
+export const UNKNOWN_JOB_STATUS = 'UNKNOWN';
